Add empty instance factory.
